streaming/requirements.txt
kafka-python==2.3.1
feedparser==6.0.11
pyspark==4.0.1
snowflake-connector-python==3.12.2
vaderSentiment==3.3.2
